ml-backend/requirements-qr.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
python-multipart==0.0.6
